device_pixel_ratio is f64

+++ src/window/glutin_window.rs
@@ -150,7 +150,7 @@ impl Window {
                         ),
                         window_width: width as usize,
                         window_height: height as usize,
-                        device_pixel_ratio: device_pixel_ratio as usize,
+                        device_pixel_ratio: device_pixel_ratio,
                         first_frame: first_frame,
                     };
                     first_frame = false;
